set -eu

+++ torture/build.sh
@@ -2,9 +2,15 @@
 #
 # TODO: This should eventually be a Github Action CI test
 
-mkdir -p z
-ln -s $(which python2) z/python
+set -eu
+
+if [ ! -e z ]; then
+	mkdir -p z
+	ln -s $(which python2) z/python
+fi
+
 export PATH="$PWD/z:$PATH" CFLAGS="-fpermissive"
+
 cd samba
 sed -i "s/HAVE_BSD_MD5_H/disabled_BSD_MD5_H/" lib/crypto/md5.h
 ./configure --disable-python \
